Updated the font sizing, and other CSS elements
I have updated the font size so that I have 3 different font weights.

I have also managed to find the fix to centring header .title in the middle of the header area, by giving the header element itself the position of relative as I was giving it to the header img element instead.